Code should be clean now.
TODO:
- clean and add tests
- rewrite documentation
- confirm all examples pass
